restructure util files. all small size. theme for every file
